Add package liveness.
Package liveness uses keepalives to determine liveness of
entities. It emits events corresponding with state changes.
(Alive -> Alive, Alive -> Dead, Dead -> Alive, Dead -> Dead)

The events that the SwitchSet emits are handled by the functions
passed to it with NewSwitch. The handlers will receive the key
of the entity, as well as the previous state of the entity.

Signed-off-by: Eric Chlebek <eric@sensu.io>